Refactor: Contexts 폴더 이동 및 경로 수정
