+++ src/components/ProductInfo/Category/index.js
@@ -8,7 +8,7 @@ import SectionBodyContent from "layout/Section/SectionBodyContent";
 
 import styles from "components/ProductInfo/ProductInfo.module.css";
 
-import { CategoryProvider } from "./Contexts";
+import { CategoryProvider } from "store/Contexts/Category";
 const Category = () => {
   return (
     <SectionBody className={styles.container}>
