+++ src/components/ProductInfo/FilterTag/SearchList.jsx
@@ -1,5 +1,5 @@
 import { useContext, useEffect, useRef, useState } from "react";
-import { ItemContext } from "./Contexts";
+import { ItemContext } from "store/Contexts/FilterTag";
 import ListItem from "./ListItem";
 
 import styles from "./SearchList.module.css";
