+++ src/components/ProductInfo/index.js
@@ -9,7 +9,7 @@ import TotalAmount from "components/ProductInfo/TotalAmount";
 import SectionWrapper from "layout/Section/SectionWrapper";
 import SectionHeader from "layout/Section/SectionHeader";
 
-import { ItemProvider } from "components/ProductInfo/FilterTag/Contexts";
+import { ItemProvider } from "store/Contexts/FilterTag";
 const ProductInfo = () => {
   return (
     <SectionWrapper>
